Feat: Added openings on join page
The "Join" page now implements the Contentful CMS to display dynamic openings and allow would-be members to read up on the requiremens and place an application.
Environment variables have been combined into a .env.example to ensure security and compliance with good practices. It is crucial to review and regenerate the keys that were exposed.